requirements.txt
PyGithub
python-dotenv
pathlib